extends basicUser and add id's
