+++ src/main/java/nl/edemtb/mtbclinicsapplication/models/BasicUser.java
@@ -7,16 +7,7 @@ import org.hibernate.annotations.Parameter;
 
 @MappedSuperclass
 abstract class BasicUser {
-    @Id
-    @GeneratedValue(generator = "sequence-generator4")
-    @GenericGenerator(name = "sequence-generator4",
-            strategy = "org.hibernate.id.enhanced.SequenceStyleGenerator",
-            parameters = {@Parameter(name = "sequence_user", value = "route_user"),
-                    @Parameter(name = "initial_value", value = "1002"),
-                    @Parameter(name = "increment_size", value = "1")
-            }
-    )
-    private Long id;
+
     @NotBlank
     @Size(min = 1, max = 20, message = "Name must be between 1-20 characters")
     private String firstName;
@@ -29,14 +20,6 @@ abstract class BasicUser {
     @NotNull
     private Integer mobileNumber;
 
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
 
     public String getFirstName() {
         return firstName;
