+++ src/main/java/nl/edemtb/mtbclinicsapplication/models/ContactForm.java
@@ -1,16 +1,24 @@
 package nl.edemtb.mtbclinicsapplication.models;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 
 @Entity
 @Table(name = "contact_forms")
 public class ContactForm extends BasicUser {
 
+    @Id
+    @GeneratedValue
+    private Long id;
     @Column(length = Integer.MAX_VALUE)
     private String message;
 
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
 
     public String getMessage() {
         return message;
