+++ src/main/java/nl/edemtb/mtbclinicsapplication/models/RegisteredUser.java
@@ -9,23 +9,13 @@ import jakarta.validation.constraints.Pattern;
 
 @Entity
 @Table(name = "registered_users")
-public class RegisteredUser {
+public class RegisteredUser extends BasicUser{
 
     @Id
     @Column(nullable = false, unique = true)
     private String username;
     @Column(nullable = false, unique = true)
     private String password;
-    @Column
-    private String firstName;
-    @Column
-    private String lastName;
-    @Column
-    @Email
-    private String email;
-    @Column
-    @Pattern(regexp = "^\\(?([+]31|0031|0)-?6(\\s?|-)([0-9]\\s{0,3}){8}$")
-    private Integer mobileNumber;
     @Column
     private String userRole;
     @Column(nullable = false)
@@ -45,38 +35,6 @@ public class RegisteredUser {
 
     public void setPassword(String password) {
         this.password = password;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public Integer getMobileNumber() {
-        return mobileNumber;
-    }
-
-    public void setMobileNumber(Integer mobileNumber) {
-        this.mobileNumber = mobileNumber;
     }
 
     public String getUserRole() {
